edited updateForm function to ref formId
