+++ package.json
@@ -10,8 +10,8 @@
   },
   "dependencies": {
     "-": "^0.0.1",
-    "@googleapis/forms": "^2.0.5",
     "@google-cloud/pubsub": "^4.3.3",
+    "@googleapis/forms": "^2.0.5",
     "canvas": "^2.11.2",
     "firebase": "^10.8.1",
     "fs": "^0.0.1-security",
